pyproject.toml
[project]
name = "grain"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "faiss-cpu==1.13.2",
  "flask==3.1.3",
  "flask-debugtoolbar==0.16.0",
  "flask-secrets==0.1.0",
  "flask-static-digest==0.4.1",
  "gunicorn==25.1.0",
  "jinja2==3.1.6",
  "minio>=7.2.20",
  "numpy==2.4.3",
  "open_clip_torch",
  "Pillow>=11.0.0",
  "psycopg==3.3.3",
  "pytest==9.0.2",
  "pytest-cov==7.0.0",
  "setuptools==82.0.1",
  "torch",
  "werkzeug==3.1.6",
]

